Altering circuit size
Switching the generic circuit size, inside
the composer from usize to u64.

+++ src/cs/composer.rs
@@ -14,7 +14,7 @@ use rand_core::{CryptoRng, RngCore};
 /// We will have a default Composer called `StandardComposer`
 pub struct StandardComposer<E: PairingEngine> {
     // n represents the number of arithmetic gates in the circuit
-    n: usize,
+    n: u64,
 
     // Selector vectors
     //
